Revert changed of holoviews.py and rename widgetlinks in links
