+++ panel/__init__.py
@@ -8,7 +8,7 @@ from . import pipeline # noqa
 from . import plotly # noqa
 from . import vega # noqa
 from . import widgets # noqa
-from . import widgetlinks # noqa
+from . import links # noqa
 
 from .interact import interact # noqa
 from .layout import Row, Column, Tabs, Spacer # noqa
